+++ core/report/generate.go
@@ -32,8 +32,7 @@ func FromManifest(m *manifest.Manifest) (report *Report, err error) {
 		return
 	}
 
-	to := time.Now()
-	from := to.Add(-oneDay)
+	to := time.Now().UTC()
 
 	var services []*Service = make([]*Service, 0)
 	report = &Report{
@@ -60,6 +59,18 @@ func FromManifest(m *manifest.Manifest) (report *Report, err error) {
 		}
 
 		for _, sl := range s.ServiceLevels {
+			duration := sl.ObservationWindow.ToDuration()
+			if duration == 0 {
+				// we use a default value, in case we did not found any
+				duration = oneDay
+			} else {
+				// We make sure to have duration rounded to 1-minute precision !!!
+				// important for AWS metrics period being multiple of 60 sec !!!
+				duration = duration.Truncate(time.Minute)
+			}
+
+			from := to.Add(-duration)
+
 			allValues := []float64{}
 			valuesHasError := false
 			for _, sli := range sl.Indicators {