+++ core/report/writer.go
@@ -10,9 +10,11 @@ import (
 
 	consolesize "github.com/nathan-fiscaletti/consolesize-go"
 	"github.com/olekukonko/tablewriter"
+	"github.com/reliablyhq/cli/core"
 	"github.com/reliablyhq/cli/core/color"
 	"github.com/reliablyhq/cli/core/iostreams"
 	"github.com/sirupsen/logrus"
+	"gopkg.in/yaml.v2"
 )
 
 /*
@@ -41,6 +43,7 @@ const (
 	TABBED     Format = "tabbed"
 	SimpleText Format = "simple"
 	MARKDOWN   Format = "markdown"
+	YAML       Format = "yaml"
 )
 
 // Write - write report based on given format
@@ -58,11 +61,15 @@ func Write(format Format, r *Report, w io.Writer, l *logrus.Logger) {
 		b, _ := json.MarshalIndent(r, "", "  ")
 		fmt.Fprintln(w, string(b))
 
+	case YAML:
+		b, _ := yaml.Marshal(r)
+		fmt.Fprintln(w, string(b))
+
 	case SimpleText:
 		reportSimpleText(r, w)
 
 	case MARKDOWN:
-		reportMarkdown(r, w)
+		_ = reportMarkdown(r, w)
 
 	default:
 		tabbedoutput(r, w)
@@ -94,7 +101,7 @@ func reportSimpleText(r *Report, w io.Writer) {
 				fmt.Fprintf(w, "%s %s: %.2f%s (last %s) [objective: %v%s, delta: %.2f%s]\n",
 					tick, sl.Name,
 					sl.Result.Actual, unit,
-					sl.ObservationWindow.To.Sub(sl.ObservationWindow.From),
+					core.HumanizeDurationShort(sl.ObservationWindow.To.Sub(sl.ObservationWindow.From)),
 					sl.Objective, unit,
 					sl.Result.Delta, unit)
 
@@ -114,10 +121,10 @@ func reportMarkdown(r *Report, w io.Writer) error {
 
 Report time: {{ dateTime .Timestamp }}
 {{ range $index, $service := .Services }}
-## Service #({{ serviceNo $index}}): {{$service.Name}}
+## Service #{{ serviceNo $index}}: {{$service.Name}}
 
-|  |  Type    | Name          | Actual | Target | Delta  |
-|--| -------- | --------------| ------ | ------ | ------ | ---------------
+|  |  Type    | Name          | Actual | Target | Delta  | Time Window  |
+|--| -------- | --------------| ------ | ------ | ------ | ------------ |
 {{ range $ind, $sl := $service.ServiceLevels }}{{ serviceLevelRow $sl }}
 {{ end }}
 {{ end }}
@@ -182,7 +189,7 @@ func markdownFuncMap() template.FuncMap {
 				fmt.Fprint(&builder, "---")
 			}
 			fmt.Fprint(&builder, "|")
-			fmt.Fprintf(&builder, "(last %s)", period)
+			fmt.Fprint(&builder, core.HumanizeDuration(period))
 			fmt.Fprint(&builder, "|")
 
 			return builder.String()
@@ -214,7 +221,9 @@ func tabbedoutput(r *Report, w io.Writer) {
 	table.SetHeader([]string{"",
 		color.Bold(color.Magenta("Actual")),
 		color.Bold(color.Magenta("Target")),
-		color.Bold(color.Magenta("Delta")), ""})
+		color.Bold(color.Magenta("Delta")),
+		color.Bold(color.Magenta("Time Window")), // ! caution: we use non-breaking space to have header not on two lines !
+	})
 
 	emptyRow := []string{"", "", "", ""}
 
@@ -233,6 +242,8 @@ func tabbedoutput(r *Report, w io.Writer) {
 			unit := "%"
 			colorFunc := color.Green
 
+			period := sl.ObservationWindow.To.Sub(sl.ObservationWindow.From)
+
 			if sl.Result == nil {
 				tick = iconUnknown
 
@@ -240,7 +251,9 @@ func tabbedoutput(r *Report, w io.Writer) {
 					fmt.Sprintf("%s %s", tick, sl.Name),
 					"---",
 					fmt.Sprintf("%v%s", sl.Objective, unit),
-					"---"}
+					"---",
+					core.HumanizeDuration(period),
+				}
 				table.Append(row)
 
 			} else {
@@ -254,7 +267,9 @@ func tabbedoutput(r *Report, w io.Writer) {
 					fmt.Sprintf("%s %s", tick, sl.Name),
 					color.Bold(colorFunc(fmt.Sprintf("%.2f%s", sl.Result.Actual, unit))),
 					fmt.Sprintf("%v%s", sl.Objective, unit),
-					fmt.Sprintf("%.2f%s", sl.Result.Delta, unit)}
+					fmt.Sprintf("%.2f%s", sl.Result.Delta, unit),
+					core.HumanizeDuration(period),
+				}
 				table.Append(row)
 
 			}